rag_requirements_TEST.txt
openai
langchain
langchain-core    # ✅ updated to meet langchain’s requirement
langchain-community
langchain-openai
langsmith
langchain-community
langgraph
# langgraph
supabase
fastapi
uvicorn
python-dotenv
azure-search-documents
azure-identity
psycopg2-binary
langgraph-checkpoint-postgres